+++ apairo/dataset/__init__.py
@@ -1,4 +1,4 @@
-from .kitti import AsyncLayoutDataset, KittiDataset
+from .kitti import AsyncLayoutDataset
 from .raw import RawDataset
 from .tartan_kitti import TartanKittiDataset
 from .goose import Goose3DDataset
@@ -33,7 +33,6 @@ def split_sequences(
 
 __all__ = [
     "AsyncLayoutDataset",
-    "KittiDataset",
     "RawDataset",
     "TartanKittiDataset",
     "Goose3DDataset",